Embed and analyze text

The embedR package provides functions for detecting SDGs in text, as well as for analyzing and visualization the hits found in text. The following provides a brief overview of the contents of the package.

Detect functions

detect_sdg detects SDGs in text using up to five different query systems: Aurora, Elsevier, SIRIS, SDSN, and OSDG

detect_any detects SDGs in text using self-specified queries utilizing the lucene-style syntax of the corpustools package.

Analysis functions

plot_sdg visualizes the relative frequency of SDG hits across query systems.

crosstab_sdg calculates cross tables of correlations between either the query systems or the different SDGs.

Datasets

projects contain random selection of research project descriptions from the P3 database of the Swiss National Science Foundation.

aurora_queries, elsevier_queries, siris_queries, sdsn_queries, auckland_queries and sdgo_queries contain a mapping of SDGs and search queries as they are employed in the respective systems.

Examples

# \donttest{
# detect SDGs using default systems
hits <- detect_sdg_systems(projects)
#> Error in detect_sdg_systems(projects): could not find function "detect_sdg_systems"

#' # detect SDGs using all five systems
hits <- detect_sdg_systems(projects,
  system = c("Aurora", "Elsevier", "SIRIS", "SDSN", "SDGO")
)
#> Error in detect_sdg_systems(projects, system = c("Aurora", "Elsevier",     "SIRIS", "SDSN", "SDGO")): could not find function "detect_sdg_systems"

# visualize SDG frequencies
plot_sdg(hits)
#> Error in plot_sdg(hits): could not find function "plot_sdg"

# correlations between systems
crosstab_sdg(hits)
#> Error in crosstab_sdg(hits): could not find function "crosstab_sdg"

# correlations between SDGs
crosstab_sdg(hits, compare = "sdgs")
#> Error in crosstab_sdg(hits, compare = "sdgs"): could not find function "crosstab_sdg"
# }